# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FAQ API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/control/settings/load": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Display a listing of the resource",
                "operationId": "16a98c8ef24e6016578c38f62b816030",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SettingsResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": []
                    },
                    {
                        "SanctumBearerToken": []
                    }
                ]
            }
        },
        "/api/control/settings/list": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "List all settings",
                "operationId": "cb1d9efa22869f7c6cc3d26f5190db66",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SettingsListResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": []
                    },
                    {
                        "SanctumBearerToken": []
                    }
                ]
            }
        },
        "/api/control/settings/show/{key}": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Show a specific setting",
                "operationId": "f143ba97b16b2014d603157c76846014",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SettingResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": []
                    },
                    {
                        "SanctumBearerToken": []
                    }
                ]
            }
        },
        "/api/control/settings/add": {
            "post": {
                "tags": [
                    "Settings"
                ],
                "summary": "Store a newly created resource in storage",
                "operationId": "8a1bfb94bee32ba218156717efee1fde",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SettingStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": []
                    },
                    {
                        "SanctumBearerToken": []
                    }
                ]
            }
        },
        "/api/control/settings/update/{key}": {
            "post": {
                "tags": [
                    "Settings"
                ],
                "summary": "Update a specific setting",
                "operationId": "3d07b5145546bb92c05fc8b41d435a40",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SettingUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resource updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": []
                    },
                    {
                        "SanctumBearerToken": []
                    }
                ]
            }
        },
        "/api/control/settings/delete/{key}": {
            "delete": {
                "tags": [
                    "Settings"
                ],
                "summary": "Remove the specified resource from storage",
                "operationId": "ad3dd1d27dd9e1e17802d1feb10537ed",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": []
                    },
                    {
                        "SanctumBearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SettingStoreRequest": {
                "title": "Setting Store Request",
                "description": "Request body for storing a new setting",
                "required": [
                    "key",
                    "value"
                ],
                "properties": {
                    "key": {
                        "type": "string",
                        "maxLength": 150,
                        "example": "site_name"
                    },
                    "value": {
                        "type": "string",
                        "maxLength": 1000,
                        "example": "My Website"
                    }
                },
                "type": "object"
            },
            "SettingUpdateRequest": {
                "title": "Setting Update Request",
                "description": "Request body for updating a setting",
                "required": [
                    "value"
                ],
                "properties": {
                    "value": {
                        "type": "string",
                        "maxLength": 1000,
                        "example": "Updated Value"
                    }
                },
                "type": "object"
            },
...
            "SettingResource": {
                "title": "Setting Resource",
                "description": "Resource representation of a setting",
                "properties": {
                    "key": {
                        "description": "The key of the setting",
                        "type": "string"
                    },
                    "value": {
                        "description": "The value of the setting",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "SettingsListResource": {
                "title": "Settings List Resource",
                "description": "Resource representation of a list of settings",
                "properties": {
                    "key": {
                        "description": "The key of the setting",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "SettingsResource": {
                "title": "Settings Resource",
                "description": "Resource representation of a setting",
                "properties": {
                    "key": {
                        "description": "The key of the setting",
                        "type": "string"
                    },
                    "value": {
                        "description": "The value of the setting",
                        "type": "string"
                    },
                    "created_user": {
                        "description": "The username of the user who created the setting",
                        "type": "string"
                    },
                    "created_date": {
                        "description": "The date and time when the setting was created",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "GeneralResource": {
                "title": "General Resource",
                "description": "General resource representation",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    },
                    "data": {
                        "description": "The data returned by the operation",
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiToken": {
                "type": "apiKey",
                "description": "ApiToken required for all end points",
                "name": "Token",
                "in": "header"
            },
            "SanctumBearerToken": {
                "type": "http",
                "description": "Sanctum Bearer Token required after user logs in (Authorization: Bearer <token>)",
                "scheme": "bearer"
            }
        }
    },
...
}